+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added SpectreConsole for a better CLI experience
 - Fixed double error messages
+- Added parallel processing for each bounded context
 
 ## [0.9.3] - 2021-04-10